refactor : PurchaseMoney 자료형 변경, GameResult의 역할 추가

+++ src/main/java/domain/PurchaseMoney.java
@@ -6,12 +6,12 @@ public class PurchaseMoney extends Money {
 
 	private static final int LOTTO_PRICE = 1_000;
 
-	public PurchaseMoney(int money) {
+	public PurchaseMoney(double money) {
 		super(money);
 		validate(money);
 	}
 
-	private void validate(int money) {
+	private void validate(double money) {
 		if (money < LOTTO_PRICE) {
 			throw new LackOfMoneyException("로또 비용이 부족합니다.");
 		}
